Changed IEntity.id IVersionSupport.version types both to native long for better overall compatability with various datastores. Note, client side (Model) still handles ids as Strings.

+++ branches/jdo2/client/persist/src/main/java/com/tll/server/rpc/entity/VersionMismatchException.java
@@ -22,7 +22,7 @@ public class VersionMismatchException extends ApplicationException {
 	 * @param entityVersion
 	 * @param modelVersion
 	 */
-	public VersionMismatchException(Class<? extends IEntity> entityClass, Integer entityVersion, Integer modelVersion) {
+	public VersionMismatchException(Class<? extends IEntity> entityClass, long entityVersion, long modelVersion) {
 		super("Version mis-match [model version: %0] [entity version: %1]", new Object[] {
 			modelVersion, entityVersion
 		});
